+++ control-base/src/main/java/fi/nls/oskari/control/data/GetSearchResultHandler.java
@@ -7,8 +7,10 @@ import fi.nls.oskari.control.ActionException;
 import fi.nls.oskari.control.ActionHandler;
 import fi.nls.oskari.control.ActionParameters;
 import fi.nls.oskari.control.ActionParamsException;
+import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
 import fi.nls.oskari.util.ResponseHelper;
+import org.json.JSONArray;
 import org.json.JSONObject;
 
 import java.util.Locale;
@@ -18,6 +20,7 @@ public class GetSearchResultHandler extends ActionHandler {
 
     private static final String PARAM_SEARCH_KEY = "searchKey";
     private static final String PARAM_EPSG_KEY = "epsg";
+    private static final String PARAM_CHANNELIDS_KEY = "channels";
 
     private String[] channels = new String[0];
 
@@ -38,20 +41,26 @@ public class GetSearchResultHandler extends ActionHandler {
         if (!SearchWorker.STR_TRUE.equals(error)) {
             // write error message key
             ResponseHelper.writeResponse(params, error);
-        } else {
-            final Locale locale = params.getLocale();
-
-            final SearchCriteria sc = new SearchCriteria(params.getUser());
-            sc.setSearchString(search);
-            sc.setSRS(epsg);  // eg. EPSG:3067
-
-            sc.setLocale(locale.getLanguage());
-
-            for(String channelId : channels) {
-                sc.addChannel(channelId);
-            }
-            final JSONObject result = SearchWorker.doSearch(sc);
-            ResponseHelper.writeResponse(params, result);
+            return;
         }
+        final Locale locale = params.getLocale();
+
+        final SearchCriteria sc = new SearchCriteria(params.getUser());
+        String[] channelIds = params.getHttpParam(PARAM_CHANNELIDS_KEY, "").split("\\s*,\\s*");
+
+        // if channels defined in request, use them
+        if(channelIds.length == 0) {
+            // otherwise use configurations, service will add defaults if left empty
+            channelIds = channels;
+        }
+        for (String id :  channelIds) {
+            sc.addChannel(id);
+        }
+        sc.setSearchString(search);
+        sc.setSRS(epsg);  // eg. EPSG:3067
+        sc.setLocale(locale.getLanguage());
+
+        final JSONObject result = SearchWorker.doSearch(sc);
+        ResponseHelper.writeResponse(params, result);
     }
 }