+++ ReleaseNotes.md
@@ -20,6 +20,22 @@ Now all available search channels that return true from SearchableChannel.isDefa
 
 IOHelper: Added a new convenience method setupBasicAuth(connection, user, pass) which sets up basic auth for the given connection.
 JSONHelper: Added a new convenience method createJSONArray(json, bln) to easily create empty arrays from null/problematic JSON param.
+content-resources/ViewHelper: Added convenience methods for easily adding a bundle to default views. 
+Flyway migrations can use them like this to add a bundle to default and user type views if the view doesn't have the bundle already:
+
+        public class Vxx_yy__add_bundle_to_views implements JdbcMigration {
+            private static final String BUNDLE_ID = "[replace with bundle id]";
+        
+            public void migrate(Connection connection) throws Exception {
+                final ArrayList<Long> views = ViewHelper.getUserAndDefaultViewIds(connection);
+                for(Long viewId : views){
+                    if (ViewHelper.viewContainsBundle(connection, BUNDLE_ID, viewId)) {
+                        continue;
+                    }
+                    ViewHelper.addBundleWithDefaults(connection, viewId, BUNDLE_ID);
+                }
+            }
+        }
 
 ### MyBatis